requirements.txt
python-dotenv>=1.0.0
Pillow>=10.0.0
schedule>=1.2.0
